refactor(core): consolidate merge algorithms
Refs #3845
